holy crap, you would think that sorting tweets in descending create_at order would be easy. but no. it was not. but it is now done
git-svn-id: http://literalshore.com/svn/kickball/trunk@529 73bdbd85-dd75-0410-9d14-8527f93b1508

+++ Classes/MGTwitterEngine.h
@@ -159,7 +159,7 @@
 
 - (NSString *)getFavoriteUpdatesFor:(NSString *)username startingAtPage:(int)pageNum; // favorites
 
-- (NSString *)markUpdate:(unsigned long)updateID asFavorite:(BOOL)flag; // favorites/create, favorites/destroy
+- (NSString *)markUpdate:(unsigned long long)updateID asFavorite:(BOOL)flag; // favorites/create, favorites/destroy
 
 
 // Notification methods
